requirements.txt
langchain>=0.2.10
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
chromadb>=0.5.0
redis>=5.0.0
fakeredis
tiktoken>=0.7.0
python-dotenv>=1.0.0